Reconstruct the code.
